Parallel running was added Now you can run tests via maven and via testng xml suite
