+++ pom.xml
@@ -19,10 +19,41 @@
         <bonigarcia.version>4.3.1</bonigarcia.version>
         <rest-assured.version>4.3.3</rest-assured.version>
         <gson.version>2.8.6</gson.version>
-        <cucumber.version>6.10.1</cucumber.version>
+        <cucumber.version>5.5.0</cucumber.version>
         <javafaker.version>1.0.2</javafaker.version>
         <hamcrest.version>2.2</hamcrest.version>
+        <maven-surefire-plugin>3.0.0-M5</maven-surefire-plugin>
+        <aspectj.version>1.9.6</aspectj.version>
     </properties>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>${maven-surefire-plugin}</version>
+                <configuration>
+                    <argLine>
+                        -javaagent:"${settings.localRepository}/org/aspectj/aspectjweaver/${aspectj.version}/aspectjweaver-${aspectj.version}.jar"
+                        -Dcucumber.options="-p pretty"
+                        -Ddataproviderthreadcount=2
+                    </argLine>
+                    <suiteXmlFiles>
+                        <suiteXmlFile>
+                            src/test/resources/testng/run_all.xml
+                        </suiteXmlFile>
+                    </suiteXmlFiles>
+                </configuration>
+                <dependencies>
+                    <dependency>
+                        <groupId>org.aspectj</groupId>
+                        <artifactId>aspectjweaver</artifactId>
+                        <version>${aspectj.version}</version>
+                    </dependency>
+                </dependencies>
+            </plugin>
+        </plugins>
+    </build>
 
     <dependencies>
         <dependency>
