+++ src/test/resources/testng/run_all.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="Qase BDD tests" parallel="tests">
+    <test name="Run all tests">
+        <classes>
+            <class name="tests.QaseTest"/>
+        </classes>
+    </test>
+</suite>
